+++ ReleaseNotes.md
@@ -1,5 +1,6 @@
 0.1.0 - 2018-??-??
 ==================
 
-- **`a_env`**: Support saving and retrieving environment variables.
-- **`a_profile`**: Look up AWS profile and access key and print the export commands to console.
+- **`arki`**: Show all tools
+- **`aws_profile`**: Look up AWS profile and access key and print the export commands to console.
+- **`env_store`**: Support saving and retrieving environment variables.